+++ src/color.rs
@@ -1,24 +1,24 @@
 use super::utils;
 use super::vec3;
 
-pub type Color = vec3::Vec3<f32>;
+pub type Color = vec3::Vec3<f64>;
 
 impl Color {
-    fn r(&self) -> f32 {
+    fn r(&self) -> f64 {
         return self.x();
     }
 
-    fn g(&self) -> f32 {
+    fn g(&self) -> f64 {
         return self.y();
     }
 
-    fn b(&self) -> f32 {
+    fn b(&self) -> f64 {
         return self.z();
     }
 }
 
 pub fn write_color(pixel_color: &Color, samples_per_pixel: &usize) -> rgb::RGBA8 {
-    let samples = *samples_per_pixel as f32;
+    let samples = *samples_per_pixel as f64;
 
     // Divide the color by the number of samples and gamma-correct for gamma=2.0.
     let r = (pixel_color.r() / samples).sqrt();
